pyproject.toml
[project]
name = "examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2.3",
    "scikit-learn>=1.7.1",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

[tool.ruff.lint]
ignore = ["E501"]
select = ["B","C","E","F","W","B9", "I", "Q"]

[tool.ruff.format]
quote-style = "double"
indent-style = "tab"
skip-magic-trailing-comma = false

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]
chapter7 = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0", 
    "textblob>=0.17.1",
]
chapter8 = [
    "hydra-core>=1.3.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]
chapter9 = [
    "loguru>=0.7.0",
]
chapter12 = [
    "pdoc3>=0.10.0",
]
chapter13 = [
    "hatchling>=1.18.0",
    "build>=1.0.0",
]
chapter14 = [
    "jupyter>=1.0.0",
    "marimo>=0.1.0",
    "ipytest>=0.13.0",
]

[tool.ruff.lint.mccabe]
max-complexity = 10
